% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to MySQL database Function}
\usage{
connect(.type = "data")
}
\arguments{
\item{.type}{(character) "data" = Cat-formatted MOVES grand data database}
}
\description{
Connect to a database of interest to this project.
}
\author{
Tim Fraser & Yan Guo
}
