% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnostics for a model based on formula}
\usage{
diagnose(.data, .formula, .pollutant, .by, .geoid)
}
\description{
For any formula, evaluate its goodness of fit and return identifiers.
}
