% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_many.R
\name{diagnose_many}
\alias{diagnose_many}
\title{Diagnostics for \strong{many} models based on formula}
\usage{
diagnose_many(
  .geoid = "36109",
  .pollutant = 98,
  .by = 8,
  .type = 42,
  .formulas = list()
)
}
\description{
For a list of formula, evaluate each model's goodness of fit and return identifiers.
}
