% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate your area-specific model}
\usage{
estimate(data, .vars = c("vmt", "vehicles", "starts", "sourcehours", "year"))
}
\arguments{
\item{.vars}{(character vector) Vector of variables to use as predictors.}

\item{.filters}{(vector/list) Named Vector or List of inputted values for filtering.}
}
\description{
Estimate your area-specific model
}
\note{
An example filter would be: \code{.filters = c(.by = 8, .pollutant = 98, .sourcetype = 42)}
}
