% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_transformation.R
\name{find_transformation}
\alias{find_transformation}
\title{find_transformation}
\usage{
find_transformation(m)
}
\arguments{
\item{m}{model object}
}
\description{
Function to find the transformation used on the model outcome, if any.
}
\author{
Tim Fraser
}
