% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{keywords}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 41 rows and 4 columns.
}
\usage{
keywords
}
\description{
Data for keywords
}
\keyword{datasets}
