% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{project}
\usage{
project(
  m,
  data,
  .newx,
  .cats = "year",
  .exclude = "geoid",
  .context = TRUE,
  .ci = 0.95
)
}
\description{
Function to generate data.frames of projected emissions given
}
