% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query}
\usage{
query(
  .db,
  .table,
  .filters = c(.by = 16, .pollutant = 98),
  .vars = c("year", "vmt", "vehicles", "starts", "sourcehours")
)
}
\arguments{
\item{.db}{database connection object for CAT GRAND database or CAT formatted MOVES output data}

\item{.table}{Name of table in database object `.db`}

\item{.filters}{Named Vector or List of inputted values for filtering.}

\item{.vars}{Vector of variables desired}
}
\description{
Query the CAT grand database for a specific slice of data.
}
\note{
.pollutant ID of the pollutant affected

.by ID of Aggregation Level (overall = `16`, by sourcetype = `8`, by fueltype = `14`, by regulatory class = `12`)

.sourcetype ID of Sourcetype

.regclass ID of Regulatory Class

.roadtype ID of Roadtype

.fueltype ID of Fueltype
}
\author{
Tim Fraser & Yan Guo
}
