% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_aggregate.R
\name{query_aggregate}
\alias{query_aggregate}
\title{query_aggregate}
\usage{
query_aggregate(data, .by = 8)
}
\arguments{
\item{.by}{Aggregation level ID to filter to. (overall = `16`, by sourcetype = `8`, by fueltype = `14`, by regulatory class = `12`)}

\item{.db}{database connection object for CAT GRAND database or CAT formatted MOVES output data}

\item{.table}{Name of table in database object `.db`}

\item{.pollutant}{ID of the pollutant affected}
}
\description{
Aggregate/relabel a queried subset, to generate mode-mix / fuel-mix
}
\author{
Tim Fraser & Yan Guo
}
