% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_many.R
\name{query_many}
\alias{query_many}
\title{query_many}
\usage{
query_many(
  .db,
  .table = "d36109",
  .filters = list(.pollutant = 98, .by = 17),
  .vars = c("vmt", "vehicles", "sourcehours", "starts")
)
}
\description{
Query multiple subsets
}
