% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setx.R
\name{setx}
\alias{setx}
\title{setx}
\usage{
setx(data, .newx, .cats = "year", .exclude = c("geoid"), .context = TRUE)
}
\arguments{
\item{data}{data.frame of default data collected from CAT Grand Database}

\item{.newx}{data.frame, named vector, or list of new values for 1 or more x variables.}

\item{.cats}{vector of stratifying variable names, for which we should get estimates for each (eg. year)}

\item{.exclude}{vector of variable names to exclude from data.frame - eg. id variables}
}
\description{
Function to create a data.frame of \code{newdata} to pass to a model, filling in the rest with default \code{data}.
}
\author{
Tim Fraser & Yan Guo
}
