% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_exists.R
\name{table_exists}
\alias{table_exists}
\title{table_exists}
\usage{
table_exists(.db, .table = "d36109")
}
\arguments{
\item{.db}{database connection object (from \code{DBI} package)}

\item{.table}{name of table potentially the supplied database \code{db}. Eg. d36109 is data for geoid 36109.}
}
\description{
Helper function to check if the supplied database table exists
}
