% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler}
\alias{tabler}
\title{TABLER}
\usage{
tabler(.geoid = "36109", .pollutant = 98, .by = 8, .type = 42)
}
\arguments{
\item{.geoid}{...}

\item{.pollutant}{...}

\item{.by}{...}

\item{.type}{...}
}
\description{
Generate a diagnostic table
}
\author{
Tim and Yan
}
