% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\docType{data}
\name{transformations}
\alias{transformations}
\title{transformations}
\format{
An object of class \code{list} of length 7.
}
\usage{
transformations
}
\description{
Data for transformation
}
\keyword{datasets}
